Decompile env_sound/toys and rename filesystem data (#101)
* Decompile env_sound and rename filesystem related data

- Add ImHex pattern for Pilotwings 64 filesystem
- Move app data symbols out of symbol_addrs_kernel.txt
- Rename functions and data associated with filesystem
  - Prefix task structs with Task* instead of Level*
  - Rename APTS and WOBJ types
  - Rename PotentialLPAD -> LevelLPAD
  - Rename func_802314D0 -> uvUserFileRead
  - Rename UVBlockOffsets::unk1838 -> ::userFiles
  - Rename gLevelCurObjects
  - Rename gLevelUserFileLookup
  - Rename Unk802E27A8_Arg0 -> LevelESND and add APTS
  - Rename WOBJ and windObj functions
  - Rename wind_objects functions with windObj prefix
  - Rename WindObjectData -> LevelWOBJ
  - Add enum for WindObjectType

- Rename and decompile functions in toys
  - Decompile toyLoad (func_80348280)
  - Decompile toyInit

- Rename code_69BF0 -> env_sound and decompile functions
  - Decompile envSoundInit (func_802E26C0)
  - Decompile envSoundLoad (func_802E27A8)
  - Decompile func_802E2904
  - Decompile func_802E2A00 (NON_MATCHING)
  - Decompile func_802E3250
  - Decompile func_802E3310
  - Decompile func_802E3398
  - Decompile envSoundFrameUpdate (func_802E344C)

* Docs: filesystem formatting and rewording

* Map env_sound .bss/.data and rename variables

* Decompile more toys and move func_803485F0 to test_menu

- Decompile toy_80347C2C
- Decompile toy_8034819C
- Decompile toy_80348214
- Decompile toy_803483AC
- Decompile toy_80348418
- Move func_803485F0 toys -> test_menu

* Decompile testMenuMainRender and cleanup test_menu interface

* Decompile toyAnimate


* Match envSound_802E2A00


---------

+++ src/app/code_7FE00.h
@@ -7,10 +7,10 @@ extern s32 D_8034F390;
 
 void func_802F8AB8(f32, f32, f32, f32, f32*);
 void func_802F8B0C(f32, f32, f32, f32);
-void func_802F8BF8(s32, s32, s32, s32, f32, f32);
+void func_802F8BF8(s32, s32, s32, f32, f32, f32);
 void func_802F9638(f32, f32, f32, f32);
 void func_802F996C(f32, f32, f32, f32);
 void func_802F9BF8(s32, f32, f32, f32, f32, f32, f32, f32, f32, f32, f32);
-s32 func_802F9D28(u8, u8, f32, f32, f32, f32, f32, f32, s32*, f32*, f32*, s32*);
+s32 func_802F9D28(u8, u8, f32, f32, f32, f32, f32, f32, f32*, f32*, f32*, s32*);
 
 #endif // APP_CODE_7FE00_H
